Update tests to match new UX
