+++ package.json
@@ -78,7 +78,7 @@
     "react-native-gesture-handler": "^1.9.0",
     "react-native-haptic-feedback": "^1.9.0",
     "react-native-icloudstore": "^0.9.0",
-    "react-native-image-picker": "^3.3.4",
+    "react-native-image-picker": "^4.0.6",
     "react-native-inappbrowser-reborn": "^3.3.4",
     "react-native-keychain": "^6.1.1",
     "react-native-localize": "^2.0.1",
